add history for automations duration is now a double (instead of int) add an "open" button in script runner panel add ClassPathHacker in order to force new jar into the classpath
git-svn-id: svn://svn.code.sf.net/p/modbuspal/code/ModbusPal@68 42ed8c19-e739-47c8-872a-484d4b446698

+++ src/modbuspal/generator/Generator.java
@@ -29,7 +29,7 @@ import org.w3c.dom.NodeList;
 public abstract class Generator
 {   
     private ImageIcon icon;
-    private int duration = 10;
+    private double duration = 10;
     private double initialValue = 0.0;
     private JPanel controlPanel;
 
@@ -113,7 +113,7 @@ public abstract class Generator
      * editorby the user of ModbusPal.
      * @return duration of the generator, in seconds.
      */
-    public int getDuration()
+    public double getDuration()
     {
         return duration;
     }
@@ -182,7 +182,7 @@ public abstract class Generator
 
         Node durNode = attributes.getNamedItem("duration");
         String durVal= durNode.getNodeValue();
-        duration = Integer.parseInt(durVal);
+        duration = Double.parseDouble(durVal);
 
         loadSettings( genNode.getChildNodes() );
     }
@@ -205,7 +205,7 @@ public abstract class Generator
      * user when he/she edits the automation.
      * @param val new duration of the generator.
      */
-    void setDuration(int val)
+    void setDuration(double val)
     {
         duration = val;
     }
